Add files via upload
Justed wanted to create an array of 52 like a card deck, however now is the challenge to make this into a deck of cards.  I'm going back to chapter 6, and going to slowly move through the book, as to get a handle of these concepts, and also I have to get better on applying what I have learned from these excerises to get a better grip on the concepts.